Add DockerFiles. Update docker-compose.yaml. Other updates to match new hosts
I managed to containerize the whole application which means that we have 7 containers: 3 for databases and 4 for the microservices including the api-gateway. Add dockerfiles to create images for the microservices and update docker-compose.yaml to containerize those images under the same network. good news: everything work fine without any problems. TODO next : try to envlope the whole containers inside of kubernetes to prepare for deployment